requirements.txt
fastapi
uvicorn
jinja2
python-multipart
langchain-openai
langchain-core
pydantic
python-dotenv
duckduckgo-search